JENKINS-36355 default timeout set to 1 second
Launching the Xvfb from docker exec needs a timeout to see if Xvfb was
launched successfully. This is due to the fact that docker exec is alive
but the nested Xvfb process launched inside the container would not be
launched already or could have crashed but docker exec itself is still
alive.