JENKINS-36355 Xvfb framebuffer directory placed in workspace
Previously Xvfb plugin would place the framebuffer in the node's root,
that is not available in docker wrapped steps. Now the framebuffer
directory is placed in the workspace directory.

Added code to creat workspace directory if it doesn't exist, in tests
workspace directory would not be created, not sure if this would be the
case outside of unit tests.